Move definitions.py into src.
