+++ src/definitions.py
@@ -1,0 +1,4 @@
+import os
+from pathlib import Path
+
+ROOT_DIR = Path(os.path.dirname(os.path.abspath(__file__))).parent
